Doing the conversion to voltage in driver

+++ Inc/Battery.hpp
@@ -6,7 +6,7 @@
 template <std::size_t N_CELLS>
 struct Battery {
     std::array<float, N_CELLS> cells{};
-    std::array<uint16_t, 4> GPIOs{};
+    std::array<float, 4> GPIOs{};
     float total_voltage{};
 };
